refactor(process): extract source test modules
- Move runner unit tests into runner_test.rs
- Keep the runtime source file focused on process code
